refactor: :recycle: refactor Board component logic into customHook
